Allow both client and server to reload weapon
