+++ crates/lumina_client/src/player/spaceship.rs
@@ -11,12 +11,10 @@ use lightyear::prelude::*;
 use lumina_common::prelude::*;
 use lumina_common::utils::ColorPalette;
 use lumina_shared::action::ReplicateActionBundle;
-use lumina_shared::game::prelude::*;
 use lumina_shared::prelude::*;
 use lumina_vfx::prelude::*;
 
 use crate::camera::MainPrepassTexture;
-use crate::effector::TeleporterEffector;
 
 use super::{CachedGameStat, LocalPlayerId};
 
@@ -39,21 +37,8 @@ impl Plugin for SpaceshipPlugin {
                     apply_heal_vfx_cooldown,
                     update_heal_vfx,
                 ),
-            )
-            .observe(teleport);
+            );
     }
-}
-
-fn teleport(
-    trigger: Trigger<TeleporterEffector>,
-    q_teleporters: Query<&Teleporter>,
-    mut connection_manager: ResMut<ConnectionManager>,
-) {
-    let Ok(&teleporter) = q_teleporters.get(trigger.entity()) else {
-        return;
-    };
-
-    let _ = connection_manager.send_message::<OrdReliableChannel, _>(&Teleport { teleporter });
 }
 
 /// Initialize the original colors of spaceship materials with the [`ShadowAbilityConfig`].
